:recycle: move the fonction getUrl for magic link into sendmail-utils
